research/pyproject.toml
[project]
name = "life-as-code-research"
version = "0.1.0"
requires-python = ">=3.12,<3.14"
dependencies = [
    "polars>=1.20",
    "duckdb>=1.2",
    "pyarrow>=18.0",
    "connectorx>=0.4",
    "scikit-learn>=1.5",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
    "pandas>=2.3.3",
    "statsmodels>=0.14.6",
]

[project.optional-dependencies]
forecasting = ["torch>=2.2,<3", "chronos-forecasting>=2.1.0", "transformers>=4.40"]
tracking = ["mlflow>=2.19"]

[project.scripts]
export-snapshot = "src.export:main"
lac-eda = "src.eda:main"
lac-preprocess = "src.preprocess:main"
lac-features = "src.features:main"
lac-baseline = "src.baseline:main"
lac-test = "src.test_engine:main"
lac-render-results = "src.render_results:main"
lac-correlations = "src.correlations:main"
lac-anomaly = "src.anomaly_tuning:main"
lac-forecast = "src.forecasting:main"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
